Removed some redundant configuration
