+++ CONTRIBUTING.md
@@ -6,4 +6,13 @@
 - Prefer relative imports (e.g., `import { fetchHelper } from '../helpers'`) over absolute imports (e.g., `import { fetchHelper } from '~/pages/page1/helpers'`).
 - Extract logic into hooks or utility functions to keep components straightforward.
 - Generally, try to make sure folders have an `index.ts` file that exports all the files that might be used from the outside.
+
+# Styling and interface design
+
 - Prefer CSS modules where possible, but don't be afraid to use semantically named global classes for straightforward cases.
+- Use sentence case for headings, button text, and other UI elements
+
+# Dumb things to avoid
+
+- Hard-coding specific colors!
+- Creating URLs by concatenating strings!
