+++ src/pages/maps/ingestion/@id/+config.h.ts
@@ -1,0 +1,10 @@
+export default {
+  meta: {
+    Page: {
+      env: {
+        client: true,
+        server: false,
+      },
+    },
+  },
+};
